Create Part I - III: Design and Simulation of a PWM Generator
Within lab part I to III a PWM-Generator has to be designed and simulated. This generator has  three subcomponents, which are addressing different learning objectives of the DHS course. In each lab one of these subcomponents is used for a practical usage of specific VHDL language constructs. In summary following content shall be learned:

1. PWM Controller: Design of a testbench in VHDL, design and implementation of test cases and simulation of designs using QuestaSim.
2. SPI: Designing of behaviour using combinational and sequential logic and implementation of a Finite State Machine (FSM).
3. PWM Driver and PWM Generator: Reusing existing components to develop complex units by hierarchical design.