Create Part IV: Implementation and Simulation of the Analog and Mixed-Signal part of the Motor Controller
The Analog/Mixed-Signal part of the motor controller is represented by a simplified model of the TI DRV 8837 chip. It contains the conversion logic and the driver circuit. In this lab you should implement only some parts of this chip to reduce complexity. The connected motor model is provided by the supervisors. Those modules operate in time-continuous domain, so it is necessary to switch the toolchain to VHDL-AMS [6]. Within the DHS lab you shall create a simulation model of the driver stage and finally combine it with the digital part to a complete AMS simulation model of the motor controller.